+++ lib/src/assists/flutter/wrap_with_expanded.dart
@@ -3,6 +3,7 @@ import 'package:analyzer_plugin/utilities/range_factory.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
 import '../../utils/ast_node_extensions.dart';
+import '../../utils/pubspec_extensions.dart';
 import '../../utils/type_checker.dart';
 
 class WrapWithExpanded extends DartAssist {
@@ -13,6 +14,8 @@ class WrapWithExpanded extends DartAssist {
     CustomLintContext context,
     SourceRange target,
   ) async {
+    if (!context.pubspec.isFlutterProject) return;
+
     context.registry.addInstanceCreationExpression((node) {
       final sourceRange = switch (node.keyword) {
         null => node.constructorName.sourceRange,