+++ lib/src/assists/flutter/wrap_with_layout_builder.dart
@@ -2,6 +2,7 @@ import 'package:analyzer/source/source_range.dart';
 import 'package:analyzer_plugin/utilities/range_factory.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
+import '../../utils/pubspec_extensions.dart';
 import '../../utils/type_checker.dart';
 
 class WrapWithLayoutBuilder extends DartAssist {
@@ -12,6 +13,8 @@ class WrapWithLayoutBuilder extends DartAssist {
     CustomLintContext context,
     SourceRange target,
   ) async {
+    if (!context.pubspec.isFlutterProject) return;
+
     context.registry.addInstanceCreationExpression((node) {
       final sourceRange = switch (node.keyword) {
         null => node.constructorName.sourceRange,