+++ lib/src/lints/flutter/avoid_returning_widgets.dart
@@ -3,6 +3,7 @@ import 'package:analyzer/error/listener.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
 import '../../utils/constants.dart';
+import '../../utils/pubspec_extensions.dart';
 import '../../utils/type_checker.dart';
 
 class AvoidReturningWidgets extends DartLintRule {
@@ -25,6 +26,8 @@ class AvoidReturningWidgets extends DartLintRule {
     ErrorReporter reporter,
     CustomLintContext context,
   ) {
+    if (!context.pubspec.isFlutterProject) return;
+
     context.registry.addMethodDeclaration((node) {
       final returnType = node.returnType?.type;
       if (returnType == null) return;