+++ lib/src/lints/flutter/prefer_border_radius_all.dart
@@ -3,6 +3,7 @@ import 'package:analyzer/error/listener.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
 import '../../utils/constants.dart';
+import '../../utils/pubspec_extensions.dart';
 import '../../utils/type_checker.dart';
 
 class PreferBorderRadiusAll extends DartLintRule {
@@ -27,6 +28,8 @@ class PreferBorderRadiusAll extends DartLintRule {
     ErrorReporter reporter,
     CustomLintContext context,
   ) {
+    if (!context.pubspec.isFlutterProject) return;
+
     context.registry.addInstanceCreationExpression((node) {
       final type = node.staticType;
       if (type == null || !borderRadiusChecker.isExactlyType(type)) return;