+++ lib/src/lints/flutter/prefer_value_changed.dart
@@ -5,6 +5,7 @@ import 'package:analyzer/error/listener.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
 import '../../utils/constants.dart';
+import '../../utils/pubspec_extensions.dart';
 
 class PreferValueChanged extends DartLintRule {
   const PreferValueChanged() : super(code: _code);
@@ -26,6 +27,8 @@ class PreferValueChanged extends DartLintRule {
     ErrorReporter reporter,
     CustomLintContext context,
   ) {
+    if (!context.pubspec.isFlutterProject) return;
+
     context.registry.addGenericFunctionType((node) {
       final returnType = node.returnType?.type;
       if (returnType is! VoidType) return;