+++ lib/src/lints/flutter/prefer_void_callback.dart
@@ -4,6 +4,7 @@ import 'package:analyzer/error/listener.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
 
 import '../../utils/constants.dart';
+import '../../utils/pubspec_extensions.dart';
 
 class PreferVoidCallback extends DartLintRule {
   const PreferVoidCallback() : super(code: _code);
@@ -24,6 +25,8 @@ class PreferVoidCallback extends DartLintRule {
     ErrorReporter reporter,
     CustomLintContext context,
   ) {
+    if (!context.pubspec.isFlutterProject) return;
+
     context.registry.addGenericFunctionType((node) {
       final returnType = node.returnType?.type;
       if (returnType is! VoidType) return;