+++ lib/src/lints/flutter/proper_edge_insets_constructor.dart
@@ -5,6 +5,7 @@ import 'package:custom_lint_builder/custom_lint_builder.dart';
 
 import '../../utils/argument_list_extensions.dart';
 import '../../utils/constants.dart';
+import '../../utils/pubspec_extensions.dart';
 import '../../utils/type_checker.dart';
 import '../../utils/utils.dart';
 
@@ -27,6 +28,8 @@ class ProperEdgeInsetsConstructor extends DartLintRule {
     ErrorReporter reporter,
     CustomLintContext context,
   ) {
+    if (!context.pubspec.isFlutterProject) return;
+
     context.registry.addInstanceCreationExpression((node) {
       final type = node.staticType;
       if (type == null || !edgeInsetsChecker.isExactlyType(type)) return;