+++ lib/src/lints/flutter/proper_expanded_and_flexible.dart
@@ -4,6 +4,7 @@ import 'package:custom_lint_builder/custom_lint_builder.dart';
 
 import '../../utils/ast_node_extensions.dart';
 import '../../utils/constants.dart';
+import '../../utils/pubspec_extensions.dart';
 import '../../utils/type_checker.dart';
 
 class ProperExpandedAndFlexible extends DartLintRule {
@@ -26,6 +27,8 @@ class ProperExpandedAndFlexible extends DartLintRule {
     ErrorReporter reporter,
     CustomLintContext context,
   ) {
+    if (!context.pubspec.isFlutterProject) return;
+
     context.registry.addInstanceCreationExpression((node) {
       final type = node.staticType;
       if (type == null ||