+++ lib/src/lints/flutter/proper_super_init_state.dart
@@ -6,6 +6,7 @@ import 'package:custom_lint_builder/custom_lint_builder.dart';
 
 import '../../utils/class_members_extensions.dart';
 import '../../utils/constants.dart';
+import '../../utils/pubspec_extensions.dart';
 import '../../utils/type_checker.dart';
 
 class ProperSuperInitState extends DartLintRule {
@@ -29,6 +30,8 @@ class ProperSuperInitState extends DartLintRule {
     ErrorReporter reporter,
     CustomLintContext context,
   ) {
+    if (!context.pubspec.isFlutterProject) return;
+
     context.registry.addClassDeclaration((node) {
       final extendsClause = node.extendsClause;
       if (extendsClause == null) return;