+++ lib/src/utils/pubspec_extensions.dart
@@ -1,0 +1,5 @@
+import 'package:pubspec_parse/pubspec_parse.dart';
+
+extension PubspecExtension on Pubspec {
+  bool get isFlutterProject => dependencies.containsKey('flutter');
+}