+++ CHANGELOG.md
@@ -9,6 +9,10 @@
   - `doc_comments_before_annotations`
   - `prefer_library_prefixes`
 
+### Changed
+
+- Disable Flutter lints and assists for Dart projects.
+
 ## [1.1.0] - 2023-10-19
 
 ### Added